refactor: 인증된 멤버정보를 member 패키지 내부에서 auth 패키지에 의존하지 않고 다룰 수 있도록 위치 변경
