+++ src/main/java/nextstep/auth/AuthConfig.java
@@ -1,7 +1,7 @@
 package nextstep.auth;
 
 import nextstep.auth.application.JwtTokenProvider;
-import nextstep.auth.ui.AuthenticationPrincipalArgumentResolver;
+import nextstep.auth.ui.LoginMemberPrincipalArgumentResolver;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
@@ -17,6 +17,6 @@ public class AuthConfig implements WebMvcConfigurer {
 
     @Override
     public void addArgumentResolvers(List argumentResolvers) {
-        argumentResolvers.add(new AuthenticationPrincipalArgumentResolver(jwtTokenProvider));
+        argumentResolvers.add(new LoginMemberPrincipalArgumentResolver(jwtTokenProvider));
     }
 }
