+++ src/main/java/nextstep/favorite/application/DeleteFavoriteService.java
@@ -2,7 +2,7 @@ package nextstep.favorite.application;
 
 import nextstep.favorite.domain.Favorite;
 import nextstep.favorite.domain.FavoriteRepository;
-import nextstep.auth.domain.LoginMember;
+import nextstep.member.domain.LoginMember;
 import nextstep.member.domain.Member;
 import nextstep.member.domain.MemberRepository;
 import org.springframework.stereotype.Service;
