+++ src/main/java/nextstep/member/domain/LoginMember.java
@@ -1,7 +1,7 @@
-package nextstep.auth.domain;
+package nextstep.member.domain;
 
 public class LoginMember {
-    private String email;
+    private final String email;
 
     public LoginMember(String email) {
         this.email = email;
