+++ src/main/java/nextstep/member/ui/LoginMemberPrincipal.java
@@ -1,4 +1,4 @@
-package nextstep.auth.ui;
+package nextstep.member.ui;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,5 +7,5 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.PARAMETER)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface AuthenticationPrincipal {
+public @interface LoginMemberPrincipal {
 }
