+++ src/main/java/nextstep/member/ui/MemberController.java
@@ -1,10 +1,9 @@
 package nextstep.member.ui;
 
-import nextstep.auth.ui.AuthenticationPrincipal;
 import nextstep.member.application.MemberService;
 import nextstep.member.application.dto.MemberRequest;
 import nextstep.member.application.dto.MemberResponse;
-import nextstep.auth.domain.LoginMember;
+import nextstep.member.domain.LoginMember;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -43,7 +42,7 @@ public class MemberController {
     }
 
     @GetMapping("/members/me")
-    public ResponseEntity<MemberResponse> findMemberOfMine(@AuthenticationPrincipal LoginMember loginMember) {
+    public ResponseEntity<MemberResponse> findMemberOfMine(@LoginMemberPrincipal LoginMember loginMember) {
         MemberResponse memberResponse = memberService.findMe(loginMember);
         return ResponseEntity.ok().body(memberResponse);
     }
